+++ src/main/java/org/orange/oie/internship2025/assetmanagementsystem/controller/AssetController.java
@@ -3,7 +3,11 @@ package org.orange.oie.internship2025.assetmanagementsystem.controller;
 import jakarta.validation.Valid;
 import org.orange.oie.internship2025.assetmanagementsystem.dto.AssetDto;
 import org.orange.oie.internship2025.assetmanagementsystem.dto.AssetRequestDto;
-import org.orange.oie.internship2025.assetmanagementsystem.service.AssetService;
+import org.orange.oie.internship2025.assetmanagementsystem.entity.AssetCategory;
+import org.orange.oie.internship2025.assetmanagementsystem.entity.AssetType;
+import org.orange.oie.internship2025.assetmanagementsystem.service.serviceInterface.AssetService;
+import org.orange.oie.internship2025.assetmanagementsystem.service.serviceInterface.AssetTypeService;
+import org.orange.oie.internship2025.assetmanagementsystem.service.serviceInterface.CategoryService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -12,13 +16,17 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 
 @RestController
-@RequestMapping("/api/assets")
+@RequestMapping("/assets")
 public class AssetController {
 
     private final AssetService assetService;
+    private final AssetTypeService typeService;
+    private final CategoryService categoryService;
 
-    public AssetController(AssetService assetService) {
+    public AssetController(AssetService assetService , AssetTypeService typeService, CategoryService categoryService) {
         this.assetService = assetService;
+        this.typeService = typeService;
+        this.categoryService = categoryService;
     }
 
     @PostMapping
@@ -33,6 +41,18 @@ public class AssetController {
     public ResponseEntity<List<AssetDto>> getAllAssets(){
         List<AssetDto> assets = assetService.getAllAssets();
         return ResponseEntity.ok(assets);
+    }
+
+    @GetMapping("/types")
+    @PreAuthorize("hasAuthority('ADMIN')")
+    public List<AssetType> getAllTypes() {
+        return typeService.getAllTypes();
+    }
+
+    @GetMapping("/categories")
+    @PreAuthorize("hasAuthority('ADMIN')")
+    public List<AssetCategory> getAllCategories() {
+        return categoryService.getAllCategories();
     }
 
 }
