--- src/main/java/org/orange/oie/internship2025/assetmanagementsystem/controller/CategoryController.java
@@ -1,27 +1,0 @@
-package org.orange.oie.internship2025.assetmanagementsystem.controller;
-
-import org.orange.oie.internship2025.assetmanagementsystem.entity.AssetCategory;
-import org.orange.oie.internship2025.assetmanagementsystem.service.CategoryService;
-import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
-
-@RestController
-@RequestMapping("/api/categories")
-public class CategoryController {
-
-    private final CategoryService categoryService;
-
-    public CategoryController(CategoryService categoryService) {
-        this.categoryService = categoryService;
-    }
-
-    @GetMapping
-    @PreAuthorize("hasAuthority('ADMIN')")
-    public List<AssetCategory> getAllCategories() {
-        return categoryService.getAllCategories();
-    }
-}
