+++ src/main/java/org/orange/oie/internship2025/assetmanagementsystem/controller/DepartmentController.java
@@ -1,7 +1,7 @@
 package org.orange.oie.internship2025.assetmanagementsystem.controller;
 
 import org.orange.oie.internship2025.assetmanagementsystem.entity.Department;
-import org.orange.oie.internship2025.assetmanagementsystem.service.DepartmentService;
+import org.orange.oie.internship2025.assetmanagementsystem.service.serviceInterface.DepartmentService;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
