+++ src/main/java/org/orange/oie/internship2025/assetmanagementsystem/controller/UserController.java
@@ -1,7 +1,7 @@
 package org.orange.oie.internship2025.assetmanagementsystem.controller;
 
 import org.orange.oie.internship2025.assetmanagementsystem.dto.UserDTO;
-import org.orange.oie.internship2025.assetmanagementsystem.service.UserService;
+import org.orange.oie.internship2025.assetmanagementsystem.service.serviceInterface.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
