+++ src/main/java/org/orange/oie/internship2025/assetmanagementsystem/entity/AssetCategory.java
@@ -11,8 +11,8 @@ import lombok.*;
 public class AssetCategory {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long categoryId;
+    private Long id;
 
-    @Column(name = "category_name", nullable = false, unique = true)
-    private String categoryName;
+    @Column(name = "name", nullable = false, unique = true)
+    private String name;
 }
