+++ src/main/java/org/orange/oie/internship2025/assetmanagementsystem/entity/AssetStatus.java
@@ -1,8 +1,8 @@
 package org.orange.oie.internship2025.assetmanagementsystem.entity;
 
 public enum AssetStatus {
-    Available,
-    Assigned,
+    AVAILABLE,
+    ASSIGNED,
     UNDER_MAINTENANCE,
     RETIRED
 }
