+++ src/main/java/org/orange/oie/internship2025/assetmanagementsystem/entity/RequestStatus.java
@@ -1,7 +1,7 @@
 package org.orange.oie.internship2025.assetmanagementsystem.entity;
 
 public enum RequestStatus {
-    Pending,
-    Accepted,
-    Rejected,
+    PENDING,
+    ACCEPTED,
+    REJECTED,
 }
