+++ src/main/java/org/orange/oie/internship2025/assetmanagementsystem/repository/UserRepository.java
@@ -12,5 +12,5 @@ import java.util.List;
 
 public interface UserRepository extends JpaRepository<User, Long>, JpaSpecificationExecutor<User> {
     User findByEmail(String email);
-    Page<User> findByDepartment_DepartmentId(Long departmentId, Pageable pageable);
+    Page<User> findByDepartment_Id(Long departmentId, Pageable pageable);
 }
