--- src/main/java/org/orange/oie/internship2025/assetmanagementsystem/service/UserService.java
@@ -1,18 +1,0 @@
-package org.orange.oie.internship2025.assetmanagementsystem.service;
-
-
-import org.orange.oie.internship2025.assetmanagementsystem.dto.UserDTO;
-import org.orange.oie.internship2025.assetmanagementsystem.entity.User;
-import org.orange.oie.internship2025.assetmanagementsystem.mapper.UserMapper;
-import org.orange.oie.internship2025.assetmanagementsystem.repository.UserRepository;
-import org.orange.oie.internship2025.assetmanagementsystem.specification.UserSpecifications;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.domain.Specification;
-import org.springframework.stereotype.Service;
-
-
-public interface UserService {
-    Page<UserDTO> searchUsers(String searchWord, String role, Long departmentId, Pageable pageable);
-}
