+++ src/main/java/org/orange/oie/internship2025/assetmanagementsystem/service/serviceImpl/AssetTypeServiceImpl.java
@@ -1,7 +1,8 @@
-package org.orange.oie.internship2025.assetmanagementsystem.service;
+package org.orange.oie.internship2025.assetmanagementsystem.service.serviceImpl;
 
 import org.orange.oie.internship2025.assetmanagementsystem.entity.AssetType;
 import org.orange.oie.internship2025.assetmanagementsystem.repository.TypeRepository;
+import org.orange.oie.internship2025.assetmanagementsystem.service.serviceInterface.AssetTypeService;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
