+++ src/main/java/org/orange/oie/internship2025/assetmanagementsystem/service/serviceInterface/AssetTypeService.java
@@ -1,4 +1,4 @@
-package org.orange.oie.internship2025.assetmanagementsystem.service;
+package org.orange.oie.internship2025.assetmanagementsystem.service.serviceInterface;
 
 import org.orange.oie.internship2025.assetmanagementsystem.entity.AssetType;
