+++ src/main/java/org/orange/oie/internship2025/assetmanagementsystem/service/serviceInterface/DepartmentService.java
@@ -1,0 +1,9 @@
+package org.orange.oie.internship2025.assetmanagementsystem.service.serviceInterface;
+
+import org.orange.oie.internship2025.assetmanagementsystem.entity.Department;
+
+import java.util.List;
+
+public interface DepartmentService {
+     List<Department> getAllDepartments();
+}
