+++ src/test/java/org/orange/oie/internship2025/assetmanagementsystem/AppTest.java
@@ -1,15 +1,15 @@
-package org.orange.oie.internship2025.assetmanagementsystem;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
-
-@SpringBootTest
-@ActiveProfiles("test")
-class AppTest {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
+//package org.orange.oie.internship2025.assetmanagementsystem;
+//
+//import org.junit.jupiter.api.Test;
+//import org.springframework.boot.test.context.SpringBootTest;
+//import org.springframework.test.context.ActiveProfiles;
+//
+//@SpringBootTest
+//@ActiveProfiles("test")
+//class AppTest {
+//
+//    @Test
+//    void contextLoads() {
+//    }
+//
+//}
